+++ src/app/Http/Controllers/PermissionController.php
@@ -2,42 +2,61 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\PermissionRequest;
 use App\Models\Permission;
-use Illuminate\Http\Request;
 
 class PermissionController extends Controller
 {
     public function index()
     {
-        $permissions = Permission::all();
+      $this->authorize('permissions');
+        $permissions = Permission::whereNotNull('id')->get();
         return view('content.permissions', compact('permissions'));
     }
 
-    public function store(Request $request)
+    public function store(PermissionRequest $request)
     {
-        $request->validate([
-            'name' => 'required|string|max:255|unique:permissions',
-        ]);
+      $this->authorize('add-permission');
+      try{
+        $validated = $request->validated();
 
-        Permission::create(['name' => $request->name]);
+        Permission::create(['name' => $validated['name']]);
 
-        return redirect()->route('permissions.index');
+        return redirect()->back()->with('success','La Permission est bien ajouter.');
+      } catch (\Exception $e) {
+        \Illuminate\Support\Facades\Log::error($e);
+
+        return redirect()->back()->with('error', 'Un Problem est survenue: ' . $e->getMessage());
+      }
     }
 
-    public function update(Request $request,  $id)
+    public function update(PermissionRequest $request, $id)
     {
-        $request->validate([
-            'name' => 'required|string|max:255|unique:permissions,name,' . $id,
-        ]);
+      $this->authorize('edit-permission');
+      try{
+        $validated = $request->validated();
 
-        Permission::where('id',$id)->update(['name' => $request->name]);
+        Permission::where('id', $id)->update(['name' => $validated['name']]);
 
-        return redirect()->route('permissions.index');
+        return redirect()->back()->with('success','La Permission est bien modifier.');
+      } catch (\Exception $e) {
+        \Illuminate\Support\Facades\Log::error($e);
+
+        return redirect()->back()->with('error', 'Un Problem est survenue: ' . $e->getMessage());
+      }
     }
 
     public function destroy($id)
     {
-        Permission::where('id',$id)->delete();
-        return redirect()->route('permissions.index');
+      $this->authorize('delete-permission');
+      try{
+        Permission::where('id', $id)->delete();
+
+        return redirect()->route('permissions.index')->with('success','La Permission est bien supprimer');
+      } catch (\Exception $e) {
+        \Illuminate\Support\Facades\Log::error($e);
+
+        return redirect()->back()->with('error', 'Un Problem est survenue: ' . $e->getMessage());
+    }
     }
 }
